Fix Core i686 tests: BLAS axpby ULP and Int width (#420)
axpby! forwards to LinearAlgebra on the parent array. Comparing that
result to broadcast `α .* x .+ β .* y` fails by 1 ULP on i686 OpenBLAS.
Compare against axpy!/axpby! on getdata instead.

Missing/nothing construction used Union{Int64,...} for integer literals,
which are Int32 on 32-bit. Use Int.

Fail-before (i686 1.12.7): axpy_axpby_tests.jl:30 Test Failed
(`2.2218...9782` vs `...978`). After: Core 14/14 axpy, 69/69 construction,
full GROUP=Core passed.



Agent-Harness: Grok Build TUI
Agent-Model: grok-4.6
Agent-Session: local session 01a07c62-9594-75e1-9f42-3faabb3d6cf2 (transcript: /home/crackauc/.grok/sessions/%2Fhome%2Fcrackauc%2Fsandbox%2Ftmp_20260907_120012_92362/01a07c62-9594-75e1-9f42-3faabb3d6cf2)

+++ test/construction_tests.jl
@@ -55,10 +55,10 @@ temp3 = ComponentArray(temp2; e = (a = 20, b = [2 4; 1 4]))
 
 # Issue #18
 temp_miss = ComponentArray(a = missing, b = [2, 1, 4, 5], c = [1, 2, 3])
-@test eltype(temp_miss) == Union{Int64, Missing}
+@test eltype(temp_miss) == Union{Int, Missing}
 @test temp_miss.a === missing
 temp_noth = ComponentArray(a = nothing, b = [2, 1, 4, 5], c = [1, 2, 3])
-@test eltype(temp_noth) == Union{Int64, Nothing}
+@test eltype(temp_noth) == Union{Int, Nothing}
 @test temp_noth.a === nothing
 
 # Issue #61
